Fix `match arms have incompatible types`
by boxing `Api` trait

+ remove `ApiCoin` type bounding
+ move `to_coin` to implementations of structs (not part of `Api` traits
anymore)
+ Fix type of `market_cap` (thx clippy :) )
+ remove unused imports (Sunday cosmetic :) )

+++ src/app/types.rs
@@ -13,7 +13,7 @@ pub struct Coin {
     pub symbol: String,
     pub quote: Option<f32>,
     pub percent_change_24h: Option<f32>,
-    pub market_cap: Option<f32>,
+    pub market_cap: Option<f64>,
 }
 
 pub type CoinList = Vec<Coin>;
